+++ apps/wfo-ui/pages/subscriptions/index.tsx
@@ -12,61 +12,14 @@ import {
     useDataDisplayParams,
 } from '@orchestrator-ui/orchestrator-ui-components';
 import { EuiPageHeader, EuiSpacer } from '@elastic/eui';
-import {
-    Subscription,
-    Subscriptions,
-} from '../../components/Subscriptions/Subscriptions';
+import { Subscriptions } from '../../components/Subscriptions/Subscriptions';
 import NoSSR from 'react-no-ssr';
 import { SUBSCRIPTIONS_TABLE_LOCAL_STORAGE_KEY } from '../../constants';
-import { WFOFilterTab } from '@orchestrator-ui/orchestrator-ui-components/src';
-
-// Todo: consider to move this out of this component
-export enum WFOSubscriptionsTabType {
-    ACTIVE = 'ACTIVE',
-    TERMINATED = 'TERMINATED',
-    TRANSIENT = 'TRANSIENT',
-    ALL = 'ALL',
-}
-
-export const defaultSubscriptionsTabs: WFOFilterTab<
+import {
+    defaultSubscriptionsTabs,
+    SubscriptionListItem,
     WFOSubscriptionsTabType,
-    Subscription
->[] = [
-    {
-        id: WFOSubscriptionsTabType.ACTIVE,
-        translationKey: 'active',
-        alwaysOnFilters: [
-            {
-                field: 'status',
-                value: 'active',
-            },
-        ],
-    },
-    {
-        id: WFOSubscriptionsTabType.TERMINATED,
-        translationKey: 'terminated',
-        alwaysOnFilters: [
-            {
-                field: 'status',
-                value: 'terminated',
-            },
-        ],
-    },
-    {
-        id: WFOSubscriptionsTabType.TRANSIENT,
-        translationKey: 'transient',
-        alwaysOnFilters: [
-            {
-                field: 'status',
-                value: 'initial-provisioning-migrating',
-            },
-        ],
-    },
-    {
-        id: WFOSubscriptionsTabType.ALL,
-        translationKey: 'all',
-    },
-];
+} from '@orchestrator-ui/orchestrator-ui-components/src';
 
 export default function SubscriptionsPage() {
     const router = useRouter();
@@ -75,7 +28,7 @@ export default function SubscriptionsPage() {
         getTableConfigFromLocalStorage(SUBSCRIPTIONS_TABLE_LOCAL_STORAGE_KEY)
             ?.selectedPageSize ?? DEFAULT_PAGE_SIZE;
     const { dataDisplayParams, setDataDisplayParam } =
-        useDataDisplayParams<Subscription>({
+        useDataDisplayParams<SubscriptionListItem>({
             pageSize: initialPageSize,
             sortBy: {
                 field: START_DATE,
