+++ packages/orchestrator-ui-components/src/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './metadata';
 export * from './processes';
+export * from './subscriptions';
